Incorporate changes made in Zeus Phase 4.1 to ensure testability and build process work

+++ db/migrate/20150720175827_remove_table_protocol_outcomes.rb
@@ -1,5 +1,7 @@
 class RemoveTableProtocolOutcomes < ActiveRecord::Migration
   def change
-    drop_table :protocol_outcomes
+    if ActiveRecord::Base.connection.table_exists? 'protocol_outcomes'
+      drop_table :protocol_outcomes
+    end
   end
 end
